requirements.txt
requests
python-dotenv
rich
jinja2
weasyprint
python-docx